@@ -1,16 +1,17 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion '25.0.0'
+    compileSdk 34
+    namespace "com.thebluealliance.spectrumsample"
 
     defaultConfig {
         applicationId "com.thebluealliance.spectrum"
         minSdkVersion 15
-        targetSdkVersion 23
+        targetSdkVersion 34
         versionCode 1
         versionName "1.0"
     }
+
     buildTypes {
         release {
             minifyEnabled false
@@ -19,18 +20,11 @@ android {
     }
 }
 
-repositories {
-    maven {
-        url "http://github.com/wada811/Android-Material-Design-Colors/raw/master/repository/"
-    }
-}
-
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile 'com.android.support:appcompat-v7:23.2.1'
-    compile 'com.android.support:preference-v7:23.2.1'
-    compile 'com.android.support:preference-v14:23.2.1'
-    compile 'com.android.support:design:23.2.1'
-    compile 'com.wada811:android-material-design-colors:3.0.0'
-    compile project(':spectrum')
+    implementation project(':spectrum')
+
+    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.8.0' // to resolve version clash
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'androidx.preference:preference:1.2.1'
+    implementation 'com.google.android.material:material:1.9.0'
 }
