@@ -1,5 +1,4 @@
 apply plugin: 'com.android.library'
-apply from: 'versioning.gradle'
 
 ext {
     bintrayRepo = 'maven'
@@ -15,7 +14,7 @@ ext {
     gitUrl = 'https://github.com/the-blue-alliance/spectrum.git'
 
     libraryVersion = '0.7.1'
-    libraryVersionCode = buildVersionCode(libraryVersion)
+    libraryVersionCode = 100 // buildVersionCode(libraryVersion)
 
     developerId = 'nwalters512'
     developerName = 'Nathan Walters'
@@ -27,16 +26,15 @@ ext {
 }
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion '25.0.0'
+    compileSdk 34
+    namespace "com.thebluealliance.spectrum"
 
     defaultConfig {
         minSdkVersion 15
-        targetSdkVersion 23
-        versionCode libraryVersionCode
-        versionName libraryVersion
+        targetSdkVersion 34
         consumerProguardFiles 'proguard-rules.pro'
     }
+
     buildTypes {
         release {
             minifyEnabled false
@@ -46,13 +44,9 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:support-v4:23.4.0'
-    compile 'com.android.support:appcompat-v7:23.4.0'
-    compile 'com.android.support:preference-v7:23.4.0'
-    compile 'com.android.support:support-annotations:23.4.0'
-    compile 'org.greenrobot:eventbus:3.0.0'
+    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.8.0' // to resolve version clash
+    implementation 'androidx.annotation:annotation:1.6.0'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'androidx.preference:preference:1.2.1'
+    implementation 'org.greenrobot:eventbus:3.3.1'
 }
-
-apply from: 'maven-install.gradle'
-apply from: 'bintray.gradle'
