@@ -3,8 +3,9 @@ package com.thebluealliance.spectrumsample;
 import com.thebluealliance.spectrum.SpectrumPreferenceCompat;
 
 import android.os.Bundle;
-import android.support.v7.preference.Preference;
-import android.support.v7.preference.PreferenceFragmentCompat;
+
+import androidx.preference.Preference;
+import androidx.preference.PreferenceFragmentCompat;
 
 public class PreferencesDemoFragment extends PreferenceFragmentCompat {
 
