@@ -8,8 +8,10 @@ import android.content.res.TypedArray;
 import android.graphics.Color;
 import android.os.Build;
 import android.preference.DialogPreference;
-import android.support.annotation.ArrayRes;
-import android.support.annotation.ColorInt;
+
+import androidx.annotation.ArrayRes;
+import androidx.annotation.ColorInt;
+
 import android.util.AttributeSet;
 import android.view.View;
 import android.view.ViewGroup;
