@@ -16,7 +16,9 @@ import android.graphics.drawable.GradientDrawable;
 import android.graphics.drawable.RippleDrawable;
 import android.graphics.drawable.StateListDrawable;
 import android.os.Build;
-import android.support.annotation.ColorInt;
+
+import androidx.annotation.ColorInt;
+
 import android.util.AttributeSet;
 import android.view.LayoutInflater;
 import android.view.View;
